Validate that killed player is not a terrorist - terrorists can not be killed directly (only later with voting)

+++ among-eus-core/src/main/java/com/zuehlke/amongeus/core/model/Game.java
@@ -120,6 +120,9 @@ public class Game {
         if (killerPlayer.getRole() != PlayerRole.TERRORIST) {
             throw new IllegalStateException("User with role " + killerPlayer.getRole() + " can not kill");
         }
+        if (killedPlayer.getRole() != PlayerRole.AGENT) {
+            throw new IllegalStateException("User with role " + killedPlayer.getRole() + " can not be killed.");
+        }
         if (state != GameState.GAME_RUNNING) {
             throw new IllegalStateException("Can not kill in game state " + state);
         }
